refactor: 각 페이지에서 사용하는 커스텀 훅은 해당 페이지 root 폴더로 이동
