+++ src/Page/Client/ClientMain/index.tsx
@@ -1,7 +1,7 @@
 import { useParams } from "react-router-dom";
 import { restaurantImage, takeoutImage } from "@/lib/images";
-import { useNavigateHereToGo } from "./hooks";
 import { Header, OrderingMethod, OrderingButton, Wrapper } from "./styles";
+import { useNavigateHereToGo } from "../hooks/useNavigateHereToGo";
 
 const ClientMain = () => {
   const { userId, storeId } = useParams();
